added secure in every project private route

+++ src/modules/project/project.service.ts
@@ -1,19 +1,42 @@
+import httpStatusCode from "http-status-codes";
 import { Project, Prisma } from "@prisma/client";
 import { prisma } from "../../config/db";
+import { JwtPayload } from "jsonwebtoken";
+import AppError from "../../errorHelpers/AppError";
 
 const createProject = async (
-  payload: Prisma.ProjectCreateInput
+  payload: Prisma.ProjectCreateInput,
+  decoded: JwtPayload
 ): Promise<Project> => {
+  if (!decoded.userId) {
+    throw new AppError(
+      httpStatusCode.UNAUTHORIZED,
+      "You are not able add project"
+    );
+  }
+
   const project = await prisma.project.create({
-    data: payload,
+    data: { ...payload, authorId: decoded.userId },
   });
   return project;
 };
 
 const updateProject = async (
   payload: Prisma.ProjectUpdateInput,
-  id: number
+  id: number,
+  decoded: JwtPayload
 ) => {
+  const project = await prisma.project.findUnique({
+    where: {
+      id,
+    },
+  });
+  if (project?.authorId !== decoded.userId && decoded.role !== "ADMIN") {
+    throw new AppError(
+      httpStatusCode.UNAUTHORIZED,
+      "You do not have permission to edit this project."
+    );
+  }
   const updatedProject = await prisma.project.update({
     where: {
       id,
@@ -28,6 +51,23 @@ const getAllProject = async () => {
   return projects;
 };
 
+const getMyProject = async (decoded: JwtPayload) => {
+  if (!decoded.userId) {
+    throw new AppError(
+      httpStatusCode.UNAUTHORIZED,
+      "You do not have permission to view this route."
+    );
+  }
+
+  const projects = await prisma.project.findMany({
+    where: {
+      authorId: Number(decoded.userId),
+    },
+  });
+
+  return projects;
+};
+
 const getProjectById = async (id: number) => {
   const project = await prisma.project.findUnique({
     where: { id },
@@ -35,7 +75,18 @@ const getProjectById = async (id: number) => {
   return project;
 };
 
-const deleteProjectById = async (id: number) => {
+const deleteProjectById = async (id: number, decoded: JwtPayload) => {
+  const project = await prisma.project.findUnique({
+    where: {
+      id,
+    },
+  });
+  if (project?.authorId !== decoded.userId && decoded.role !== "ADMIN") {
+    throw new AppError(
+      httpStatusCode.UNAUTHORIZED,
+      "You do not have permission to delete this project."
+    );
+  }
   await prisma.project.delete({
     where: { id },
   });
@@ -48,4 +99,5 @@ export const ProjectService = {
   getAllProject,
   getProjectById,
   deleteProjectById,
+  getMyProject,
 };
